@@ -11,6 +11,6 @@ public interface Labels_info {
         put(Labels_Keys.MOUSE , new float[]{6.5f, 263.621f});
     }};
     Map<String, Float> objectHeight = new HashMap<String, Float>() {{
-        put(Labels_Keys.PERSON, 180f);
+        put(Labels_Keys.PERSON, 1750f);
     }};
 }
